+++ cc-service-api/src/main/java/com/jtmc/apps/reforma/api/v1/branch/BranchApi.java
@@ -27,5 +27,11 @@ public interface BranchApi {
     @Path("/{branchId}")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
+    Response updateBranch(@PathParam("branchId") int branchId, BranchRequest branchRequest);
+
+    @POST
+    @Path("/{branchId}/initial-amount")
+    @Consumes(MediaType.APPLICATION_JSON)
+    @Produces(MediaType.APPLICATION_JSON)
     Response insertBranchInitialAmount(@PathParam("branchId") int branchId, BranchInitialAmount initialAmount);
 }