+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Catalog Counts Service Changelog
 
+
+## 2.4.0
+release date: Oct 13, 2024
+
+### Features
+- change initial-amount POST action to it's own api endpoint
+- add timezoneId and status to branch request
+  - this will allow basic branch/account setup
+- Persona API to allow CRUD operations
+
 ## 2.3.3
 release date: Jul 7, 2024